Refactored Project Structure to better organize objects.
