+++ DaggerDemo2/app/src/main/java/com/tc2r1/daggerdemocars/di/CarComponent.kt
@@ -4,7 +4,7 @@ import com.tc2r1.daggerdemocars.MainActivity
 import com.tc2r1.daggerdemocars.objects.Car
 import dagger.Component
 
-@Component(modules = [PetrolEngineModule::class])
+@Component(modules = [PetrolEngineModule::class, WheelsModule::class])
 interface CarComponent {
 
     fun getCar(): Car
