+++ DaggerDemo2/app/src/main/java/com/tc2r1/daggerdemocars/di/DieselEngineModule.kt
@@ -1,8 +1,7 @@
 package com.tc2r1.daggerdemocars.di
 
-import com.tc2r1.daggerdemocars.objects.DieselEngine
-import com.tc2r1.daggerdemocars.objects.Engine
-import com.tc2r1.daggerdemocars.objects.PetrolEngine
+import com.tc2r1.daggerdemocars.objects.hood.DieselEngine
+import com.tc2r1.daggerdemocars.objects.hood.Engine
 import dagger.Module
 import dagger.Provides
