+++ DaggerDemo2/app/src/main/java/com/tc2r1/daggerdemocars/di/WheelsModule.kt
@@ -1,4 +1,29 @@
 package com.tc2r1.daggerdemocars.di
 
-class Wheels {
+import com.tc2r1.daggerdemocars.objects.mobility.Rims
+import com.tc2r1.daggerdemocars.objects.mobility.Tires
+import com.tc2r1.daggerdemocars.objects.mobility.Wheels
+import dagger.Module
+import dagger.Provides
+
+@Module
+class WheelsModule {
+
+    @Provides
+    fun providesTires(): Tires {
+        println("Tires Created")
+        return Tires()
+    }
+
+    @Provides
+    fun providesRims(): Rims {
+        println("Rims Created")
+        return Rims()
+    }
+
+    @Provides
+    fun providesWheels(rims: Rims, tires: Tires): Wheels {
+        println("Wheels Created")
+        return Wheels(rims, tires)
+    }
 }
