+++ DaggerDemo2/app/src/main/java/com/tc2r1/daggerdemocars/objects/mobility/Wheels.kt
@@ -1,5 +1,3 @@
-package com.tc2r1.daggerdemocars.objects
-
-import javax.inject.Inject
+package com.tc2r1.daggerdemocars.objects.mobility
 
 class Wheels(var rims: Rims, var tires: Tires)
